Refactor channel_system module to be consistent with program_graph
