+++ Cargo.toml
@@ -56,7 +56,7 @@ members = [
 [workspace.dependencies]
 anyhow = "1.0.102"
 csv = "1.4.0"
-log = "0.4.32"
+log = "0.4.33"
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.150"
 thiserror = "2.0.18"
